+++ src/lisp/kernel/clos/package.lisp
@@ -147,3 +147,7 @@
 #+clasp
 (export '(no-applicable-method-error))
 
+#+clasp
+(export '(start-profiling stop-profiling
+          report-profiling profiling-data
+          with-profiling))